+++ lib/eve_online.rb
@@ -68,6 +68,8 @@ require 'eve_online/esi/universe_region'
 require 'eve_online/esi/universe_constellations'
 require 'eve_online/esi/universe_constellation'
 require 'eve_online/esi/universe_systems'
+require 'eve_online/esi/wars'
+require 'eve_online/esi/war_killmails'
 
 require 'eve_online/esi/models/base'
 require 'eve_online/esi/models/loyalty_point'
@@ -113,6 +115,8 @@ require 'eve_online/esi/models/faction'
 require 'eve_online/esi/models/ancestry'
 require 'eve_online/esi/models/region'
 require 'eve_online/esi/models/constellation'
+
+require 'eve_online/esi/models/killmail_short'
 
 module EveOnline
 end