+++ CHANGELOG.md
@@ -14,6 +14,9 @@
 * Add `EveOnline::ESI::UniverseConstellation`
 * Add `EveOnline::ESI::Models::Constellation`
 * Add `EveOnline::ESI::UniverseSystems`
+* Add `EveOnline::ESI::Wars`
+* Add `EveOnline::ESI::Models::KillmailShort`
+* Add `EveOnline::ESI::WarKillmails`
 
 ## v0.17.0